Changed index to more appropriate type.

+++ prototype/proto-game/src/core_api.rs
@@ -3,7 +3,7 @@ pub type RomDataPointer = *const u8;
 #[link(wasm_import_module = "gpu")]
 extern {
     #[link_name = "set_object"]
-    fn gpu_set_object(index: usize, ptr: RomDataPointer, size: usize);
+    fn gpu_set_object(index: u16, ptr: RomDataPointer, size: usize);
 }
 
 #[link(wasm_import_module = "logger")]
@@ -71,7 +71,7 @@ impl Objects {
         Self {}
     }
 
-    pub fn set(&mut self, index: usize, record: RomDataRecord) {
+    pub fn set(&mut self, index: u16, record: RomDataRecord) {
         unsafe { gpu_set_object(index, record.ptr, record.size) };
     }
 }
